# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anime Gallery Media Server",
    "description": "\nAnime Gallery Media Server API 覆盖前台图片浏览、后台媒体资料维护、批量上传任务、\n元数据导入、部署设置与系统健康检查。\n\n鉴权模型：\n\n- 浏览器后台会话使用 `agms_admin_session` HttpOnly Cookie。\n- 运维脚本建议使用 `Authorization: Bearer <AGMS_API_KEYS value>`，并只授予任务实际需要的权限。\n- 使用 Cookie 会话调用非安全方法时，还需要提交与 `agms_admin_csrf` Cookie 匹配的 `X-CSRF-Token`。\n- API Key 不依赖 Cookie，因此不需要 CSRF；可在后台独立创建、设置到期时间、调整权限或立即撤销。\n- 旧版本升级后，已有 API Key 默认迁移为全部权限，避免现有自动化任务中断。\n",
    "version": "0.1.5"
  },
  "servers": [
    {
      "url": "/",
      "description": "当前部署域名"
    },
    {
      "url": "/api",
      "description": "API 前缀"
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "管理员登录",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "examples": {
                "adminLogin": {
                  "summary": "浏览器管理员登录",
                  "value": {
                    "username": "admin",
                    "password": "use-a-strong-password-2026"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "description": "创建 HttpOnly 管理员会话 Cookie。脚本和自动化任务应优先使用 API Key，而不是模拟浏览器登录。"
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "退出当前管理员会话",
        "operationId": "logout_api_auth_logout_post",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "撤销当前 Cookie 会话并清理管理员 Cookie。",
        "x-api-key-scopes": []
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "查看当前管理员身份",
        "operationId": "me_api_auth_me_get",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUser"
                },
                "examples": {
                  "example": {
                    "summary": "示例响应",
                    "value": {
                      "username": "admin",
                      "avatar_image_id": null,
                      "avatar_image": null
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "返回已登录管理员资料。支持浏览器会话 Cookie 或运维 API Key。",
        "x-api-key-scopes": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthUser": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "avatar_image_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Image Id"
          },
          "avatar_image": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "password_change_required": {
            "type": "boolean",
            "title": "Password Change Required",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "AuthUser"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImageSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "thumbnail_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Path"
          },
          "preview_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Path"
          },
          "media_version": {
            "type": "integer",
            "title": "Media Version",
            "default": 1
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          },
          "orientation": {
            "type": "string",
            "title": "Orientation"
          },
          "rating": {
            "type": "string",
            "title": "Rating"
          },
          "favorite_count": {
            "type": "integer",
            "title": "Favorite Count"
          },
          "dynamic_range": {
            "type": "string",
            "title": "Dynamic Range"
          },
          "is_animated": {
            "type": "boolean",
            "title": "Is Animated"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "width",
          "height",
          "orientation",
          "rating",
          "favorite_count",
          "dynamic_range",
          "is_animated",
          "created_at"
        ],
        "title": "ImageSummary"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "default": ""
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "cookie"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "avatar_image_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Image Id"
          },
          "avatar_image": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "password_change_required": {
            "type": "boolean",
            "title": "Password Change Required",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "expires_in",
          "username"
        ],
        "title": "LoginResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OperationsApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "AGMS_API_KEY",
        "description": "后台创建的可分权运维 API Key。权限：`library:read` 媒体库读取；`uploads:manage` 图片上传；`library:write` 资料维护；`library:delete` 永久删除；`system:read` 系统只读；`settings:manage` 系统设置；`updates:read` 更新只读；`updates:run` 执行更新；`api_keys:manage` API Key 管理"
      },
      "AdminSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agms_admin_session",
        "description": "浏览器管理员会话 Cookie。"
      },
      "CsrfTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agms_admin_csrf",
        "description": "管理员会话同时签发的 CSRF Cookie。"
      },
      "CsrfTokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "使用 Cookie 会话调用非安全方法时，必须与 CSRF Cookie 一致。"
      }
    }
  },
...
}